Updated and cleaned rubocop setup

+++ easybib/libraries/cron.rb
@@ -14,9 +14,7 @@ module EasyBib
 
       collection = []
 
-      unless ::File.exist?(@file)
-        return collection
-      end
+      return collection unless ::File.exist?(@file)
 
       crontabs = ::File.open(@file)
